advanced-rag/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
chromadb>=0.4.0
langchain>=0.0.325
sentence-transformers>=2.2.0
pypdf>=3.0.0
numpy>=1.21.0
pandas>=1.3.0
umap-learn>=0.5.0
matplotlib>=3.4.0
typing>=3.7.4
pydantic>=1.8.2
tqdm>=4.62.0